Improvements for ImageSequence and LoadingWidget
* Added a new setting to make loading widget stay hidden until after
the movies have finished playing.

* Added DesiredSize setting for ImageSequence to set a fixed size
in pixels. This avoids issues when running in standalone mode
caused by having an incorrect original size of the image during startup.

* Added the ability to preload the sequence images in the same way as
the background images, controlled by the same setting bPreloadImages.
This was an attempt to get the right size of the images for the scale
to work but it didn't seem to help. The preloading is still worth
to keep.